+++ tests/benchmarking/test_micro_benchmark.py
@@ -11,7 +11,7 @@ def test_micro_benchmark(t_sleep: float, silent: bool):
         high_precision_sleep(t_sleep)
 
     # --- act ---------------------------------------------
-    t_est = benchmark(f_test, t_per_run=0.05, n_warmup=5, n_benchmark=10, silent=silent)
+    t_est = benchmark(f_test, t_per_run=0.1, n_warmup=5, n_benchmark=10, silent=silent)
 
     # --- assert ------------------------------------------
     assert 0.5 * t_sleep <= t_est <= 2 * t_sleep