+++ tests/benchmarking/test_sleep.py
@@ -5,20 +5,14 @@ import pytest
 from brtp.benchmarking import high_precision_sleep
 
 
-@pytest.mark.parametrize("total_duration_sec,n_chunks", [(1e-3, 100), (1e-3, 10), (1e-2, 20), (1e-1, 50), (1e-1, 1)])
-def test_high_precision_sleep(total_duration_sec: float, n_chunks: int):
-    # --- arrange -----------------------------------------
-    sleep_per_iter = total_duration_sec / n_chunks
-
+@pytest.mark.parametrize("sleep_duration_sec", [1e-4, 1e-3, 1e-2])
+def test_high_precision_sleep(sleep_duration_sec: float):
     # --- act ---------------------------------------------
     t_start = time.perf_counter()
-    for _ in range(n_chunks):
-        high_precision_sleep(sleep_per_iter)
+    high_precision_sleep(sleep_duration_sec)
     t_end = time.perf_counter()
 
-    t_total = t_end - t_start
-
-    print(t_total, total_duration_sec)
+    actual_sleep_sec = t_end - t_start
 
     # --- assert ------------------------------------------
-    assert 0.5 * total_duration_sec <= t_total <= 2.0 * total_duration_sec
+    assert 0.5 * sleep_duration_sec <= actual_sleep_sec <= 2.0 * sleep_duration_sec